Create README.md
Selling feet pictures might seem like one of the more unusual ideas in a world where making money online has become more commonplace, but it’s a growing industry. This growth has been spurred on by the emergence of niche content platforms, privacy-centric side hustles, and creator monetization — selling feet pictures is no longer as much of a taboo. But is it actually worth it?